# api_docs/swagger.yaml
swagger: '2.0'
info:
  title: Sensordata API
  description: REST API for Cubesensor data
  version: 1.0.0
host: localhost:8080
schemes:
  - http
basePath: /api
produces:
  - application/json
paths:
  /sensors:
    get:
      summary: Get all sensor ids
      responses:
        '200':
          description: List of sensor ids in the database
          schema:
            type: array
            items:
              type: string
  /status:
    get:
      summary: Get latest knows status for each sensor
      responses:
        '200':
          description: List of sensor statuses
  /data:
    get:
      summary: Get rows for all sensors
      responses:
        '200':
          description: Rows of sensor data
      parameters:
        -
          name: take
          in: query
          description: Number of rows to return
          required: false
          type: number
          format: integer
          default: 1
        -
          name: skip
          in: query
          description: Number of rows to skip
          required: false
          type: number
          format: integer
          default: 0
  /data/{sensor_id}:
    get:
      summary: Get rows for selected sensor
      responses:
        '200':
          description: Rows of sensor data
      parameters:
        -
          name: sensor_id
          in: path
          description: Id of selected sensor
          required: true
          type: string
        -
          name: take
          in: query
          description: Number of rows to return
          required: false
          type: number
          format: integer
          default: 1
        -
          name: skip
          in: query
          description: Number of rows to skip
          required: false
          type: number
          format: integer
          default: 0
  /avg/{field_id}/{sensor_id}:
    get:
      summary: Get average value for selected data
      responses:
        '200':
          description: Average value
      parameters:
        -
          name: field_id
          in: path
          description: Selected field
          required: true
          type: string
          enum: 
            - temperature
            - pressure
            - humidity
            - voc
            - light
            - noise
        -
          name: sensor_id
          in: path
          description: Id of selected sensor
          required: true
          type: string
        -
          name: minutes
          in: query
          description: Duration where average is calculated
          required: false
          type: number
          format: integer
          default: 60